feat: major upgrade - parallel sync, split dashboard, AI cache, config validation
- Fix refresh.go bug: replace broken time.Parse with strconv.Atoi for bare number intervals
- Parallelize PR sync with goroutine worker pool (5 concurrent gh calls)
- Make stale threshold configurable via config instead of hardcoded 3-day values
- Implement prflow ls command (quick list from cache, no TUI)
- Implement prflow sync command (actually syncs PRs from GitHub)
- Add morning summary header showing action counts and stale reviewer stats
- Add SQLite indexes for section+updated_at, repo+number, thread lookups
- Add AI analysis caching in SQLite with 1-hour TTL
- Split monolithic dashboard.go (2000+ lines) into 4 focused files:
  dashboard.go (model/update), dashboard_sync.go, dashboard_actions.go, dashboard_render.go
- Add config validation with Validate() and ParseStaleThresholdDays()
- Use configured merge method in merge confirmation instead of hardcoded "squash"
- Use sectionCount const for modulo navigation instead of magic number
- Remove compiled binary from git tracking
- Fix pre-existing workspace test failures (gpg signing, explicit git add)
- Add 135+ new test cases (121 -> 256 total) covering all changes
- Downgrade go.mod to 1.24.7 for broader compatibility
