+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/cheenu1092-oss/prflow
 
-go 1.25.0
+go 1.24.7
 
 require (
 	github.com/aymanbagabas/go-osc52/v2 v2.0.1 // indirect
