+++ internal/tui/dashboard.go
@@ -1,13 +1,11 @@
 package tui
 
 import (
+	"encoding/json"
 	"fmt"
-	"os"
-	"strings"
 	"time"
 
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
 
 	"github.com/cheenu1092-oss/prflow/internal/ai"
 	"github.com/cheenu1092-oss/prflow/internal/cache"
@@ -23,6 +21,7 @@ const (
 	sectionReview
 	sectionWorkspace
 	sectionNeedsAttention
+	sectionCount // used for modulo navigation
 )
 
 func (s section) String() string {
@@ -51,21 +50,21 @@ const (
 )
 
 type dashModel struct {
-	cfg       *config.Config
-	db        *cache.DB
-	username  string
+	cfg      *config.Config
+	db       *cache.DB
+	username string
 
 	// Navigation
-	section   section
-	cursor    int
-	viewMode  viewMode
+	section  section
+	cursor   int
+	viewMode viewMode
 
 	// Data
-	doNow           []cache.CachedPR
-	waiting         []cache.CachedPR
-	review          []cache.CachedPR
-	needsAttention  []cache.CachedPR
-	workspace       []RepoStatus
+	doNow          []cache.CachedPR
+	waiting        []cache.CachedPR
+	review         []cache.CachedPR
+	needsAttention []cache.CachedPR
+	workspace      []RepoStatus
 
 	// Detail view
 	detailPR      *cache.CachedPR
@@ -83,10 +82,10 @@ type dashModel struct {
 	replyThreadID string
 
 	// AI analysis
-	aiAnalysis    *ai.PRAnalysis
-	aiThread      *ai.ThreadAnalysis
-	aiLoading     bool
-	aiAvailable   bool
+	aiAnalysis  *ai.PRAnalysis
+	aiThread    *ai.ThreadAnalysis
+	aiLoading   bool
+	aiAvailable bool
 
 	// State
 	loading    bool
@@ -157,6 +156,8 @@ type prActionDoneMsg struct {
 type dashTickMsg time.Time
 
 func RunDashboard(cfg *config.Config) error {
+	cfg.Validate()
+
 	db, err := cache.Open()
 	if err != nil {
 		return fmt.Errorf("failed to open cache: %w", err)
@@ -192,280 +193,6 @@ func (m dashModel) Init() tea.Cmd {
 		dashTickCmd(),
 		startAutoRefreshTimer(m.cfg.Settings.RefreshInterval),
 	)
-}
-
-func syncPRs(db *cache.DB, cfg *config.Config, username string) tea.Cmd {
-	return func() tea.Msg {
-		var doNow, waiting, review []cache.CachedPR
-		seenPRs := make(map[string]bool)
-
-		// Re-verify username if it's the default fallback
-		if username == "" || username == "user" {
-			if u, err := gh.CheckAuth(); err == nil && u != "user" {
-				username = u
-			}
-		}
-
-		// Step 1: Search for my authored PRs (fast, cross-repo)
-		myPRs, _ := gh.SearchMyPRs()
-
-		// Build set of repos + my PR numbers from search results
-		repoSet := make(map[string]bool)
-		myPRKeys := make(map[string]bool) // "repo#number" keys for PRs I authored
-		for _, pr := range myPRs {
-			if pr.Repository.NameWithOwner != "" {
-				repoSet[pr.Repository.NameWithOwner] = true
-				myPRKeys[fmt.Sprintf("%s#%d", pr.Repository.NameWithOwner, pr.Number)] = true
-			}
-		}
-		for _, repo := range cfg.Repos {
-			repoSet[repo] = true
-		}
-
-		// Step 2: For each repo, get rich PR data
-		for repo := range repoSet {
-			repoPRs, err := gh.ListPRsForRepo(repo)
-			if err != nil {
-				continue
-			}
-			for i := range repoPRs {
-				pr := &repoPRs[i]
-				key := fmt.Sprintf("%s#%d", repo, pr.Number)
-				if seenPRs[key] {
-					continue
-				}
-				seenPRs[key] = true
-
-				// Check if this is my PR: either from search results or by username match
-				isMyPR := myPRKeys[key] || strings.EqualFold(pr.Author.Login, username)
-
-				cached := cache.CachedPR{
-					PR:   *pr,
-					Repo: repo,
-				}
-
-				if isMyPR {
-					switch {
-					case pr.ReviewDecision == "CHANGES_REQUESTED":
-						cached.Section = "do_now"
-						doNow = append(doNow, cached)
-					case pr.ReviewDecision == "APPROVED":
-						cached.Section = "do_now"
-						doNow = append(doNow, cached)
-					case pr.Mergeable == "CONFLICTING":
-						cached.Section = "do_now"
-						doNow = append(doNow, cached)
-					default:
-						cached.Section = "waiting"
-						waiting = append(waiting, cached)
-					}
-				}
-				// Non-authored PRs will be caught by review requests below
-
-				db.UpsertPR(pr, repo, cached.Section)
-			}
-		}
-
-		// Fallback: if per-repo didn't work, use search results directly
-		if len(doNow) == 0 && len(waiting) == 0 && len(myPRs) > 0 {
-			for i := range myPRs {
-				pr := &myPRs[i]
-				key := fmt.Sprintf("%s#%d", pr.Repository.NameWithOwner, pr.Number)
-				if seenPRs[key] {
-					continue
-				}
-				seenPRs[key] = true
-				cached := cache.CachedPR{
-					PR:      *pr,
-					Repo:    pr.Repository.NameWithOwner,
-					Section: "waiting",
-				}
-				waiting = append(waiting, cached)
-				db.UpsertPR(pr, cached.Repo, "waiting")
-			}
-		}
-
-		// Step 3: Get PRs where review is requested from me
-		reviewPRs, _ := gh.SearchReviewRequests()
-		for i := range reviewPRs {
-			pr := &reviewPRs[i]
-			key := fmt.Sprintf("%s#%d", pr.Repository.NameWithOwner, pr.Number)
-			if seenPRs[key] {
-				continue
-			}
-			seenPRs[key] = true
-			cached := cache.CachedPR{
-				PR:      *pr,
-				Repo:    pr.Repository.NameWithOwner,
-				Section: "review",
-			}
-			review = append(review, cached)
-			db.UpsertPR(pr, cached.Repo, "review")
-		}
-
-		// Step 4: Get PRs needing re-attention (reviewed by me, but updated after my review)
-		var needsAttention []cache.CachedPR
-		reviewedPRs, _ := gh.SearchReviewedPRs()
-		for i := range reviewedPRs {
-			pr := &reviewedPRs[i]
-			key := fmt.Sprintf("%s#%d", pr.Repository.NameWithOwner, pr.Number)
-			
-			// Skip if already in another section (e.g., my own PRs or review requests)
-			if seenPRs[key] {
-				continue
-			}
-			
-			// Get full PR details to check if it needs attention
-			detail, err := gh.GetPRDetail(pr.Repository.NameWithOwner, pr.Number)
-			if err != nil {
-				continue
-			}
-			
-			// Check if there's been activity after my last review
-			if needsReReview(detail, username) {
-				seenPRs[key] = true
-				cached := cache.CachedPR{
-					PR:      *detail,
-					Repo:    pr.Repository.NameWithOwner,
-					Section: "needs_attention",
-				}
-				needsAttention = append(needsAttention, cached)
-				db.UpsertPR(detail, cached.Repo, "needs_attention")
-			}
-		}
-
-		// Step 5: Sort each section by urgency (not just updated_at)
-		SortByUrgency(doNow)
-		SortByUrgency(waiting)
-		SortByUrgency(review)
-		SortByUrgency(needsAttention)
-
-		return syncDoneMsg{
-			doNow:          doNow,
-			waiting:        waiting,
-			review:         review,
-			needsAttention: needsAttention,
-		}
-	}
-}
-
-// needsReReview checks if a PR needs re-attention from the reviewer
-// Returns true if the PR was updated after the user's last review
-func needsReReview(pr *gh.PR, username string) bool {
-	if pr == nil || username == "" {
-		return false
-	}
-	
-	// Skip if this is my own PR
-	if strings.EqualFold(pr.Author.Login, username) {
-		return false
-	}
-	
-	// Find my last review timestamp
-	var myLastReview time.Time
-	for _, rev := range pr.Reviews.Nodes {
-		if !strings.EqualFold(rev.Author.Login, username) {
-			continue
-		}
-		
-		reviewTime, err := time.Parse(time.RFC3339, rev.SubmittedAt)
-		if err != nil {
-			continue
-		}
-		
-		if reviewTime.After(myLastReview) {
-			myLastReview = reviewTime
-		}
-	}
-	
-	// If I never reviewed, no need for re-attention
-	if myLastReview.IsZero() {
-		return false
-	}
-	
-	// Check if PR was updated after my last review
-	prUpdated, err := time.Parse(time.RFC3339, pr.UpdatedAt)
-	if err != nil {
-		return false
-	}
-	
-	// If PR updated after my review (with 1-minute buffer to avoid false positives)
-	return prUpdated.After(myLastReview.Add(1 * time.Minute))
-}
-
-func scanWorkspace(cfg *config.Config) tea.Cmd {
-	return func() tea.Msg {
-		var repos []RepoStatus
-		for _, dir := range cfg.Workspace.ScanDirs {
-			// Scan for git repos in directory
-			entries, err := scanDir(dir)
-			if err != nil {
-				continue
-			}
-			for _, path := range entries {
-				rs, err := ScanWorkspaceRepo(path)
-				if err != nil {
-					continue
-				}
-				repos = append(repos, *rs)
-			}
-		}
-		// Also scan explicitly configured repos
-		for _, path := range cfg.Workspace.Repos {
-			rs, err := ScanWorkspaceRepo(path)
-			if err != nil {
-				continue
-			}
-			repos = append(repos, *rs)
-		}
-		return workspaceScanMsg{repos: repos}
-	}
-}
-
-func scanDir(dir string) ([]string, error) {
-	entries, err := readDirNames(dir)
-	if err != nil {
-		return nil, err
-	}
-	var repos []string
-	for _, name := range entries {
-		path := dir + "/" + name
-		if isGitRepo(path) {
-			repos = append(repos, path)
-		} else {
-			// Scan 2 levels deep (e.g., ~/repos/org/repo)
-			subEntries, err := readDirNames(path)
-			if err != nil {
-				continue
-			}
-			for _, subName := range subEntries {
-				subPath := path + "/" + subName
-				if isGitRepo(subPath) {
-					repos = append(repos, subPath)
-				}
-			}
-		}
-	}
-	return repos, nil
-}
-
-func readDirNames(dir string) ([]string, error) {
-	entries, err := os.ReadDir(dir)
-	if err != nil {
-		return nil, err
-	}
-	var names []string
-	for _, e := range entries {
-		if e.IsDir() && !strings.HasPrefix(e.Name(), ".") {
-			names = append(names, e.Name())
-		}
-	}
-	return names, nil
-}
-
-func isGitRepo(path string) bool {
-	_, err := gitCmd(path, "rev-parse", "--is-inside-work-tree")
-	return err == nil
 }
 
 func (m dashModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
@@ -513,11 +240,11 @@ func (m dashModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 				return m, nil
 			}
 		case "tab":
-			m.section = (m.section + 1) % 5
+			m.section = (m.section + 1) % sectionCount
 			m.cursor = 0
 			m.viewMode = viewList
 		case "shift+tab":
-			m.section = (m.section + 4) % 5
+			m.section = (m.section + sectionCount - 1) % sectionCount
 			m.cursor = 0
 			m.viewMode = viewList
 		case "up", "k":
@@ -585,18 +312,19 @@ func (m dashModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 					m.statusMsg = "✗ Cannot merge draft PR"
 					return m, nil
 				}
-				// Simple confirmation: user must press 'm' twice in a row
-				// If statusMsg contains "Press 'm' again to confirm", proceed
-				// Otherwise, set confirmation message
-				if m.statusMsg == "⚠️ Press 'm' again to confirm merge (squash)" {
+				mergeMethod := m.cfg.Settings.MergeMethod
+				if mergeMethod == "" {
+					mergeMethod = "squash"
+				}
+				confirmMsg := fmt.Sprintf("⚠️ Press 'm' again to confirm merge (%s)", mergeMethod)
+				if m.statusMsg == confirmMsg {
 					m.statusMsg = "Merging..."
-					// Default to squash merge (most common)
 					return m, func() tea.Msg {
-						err := gh.MergePR(pr.Repo, pr.Number, "squash", false)
+						err := gh.MergePR(pr.Repo, pr.Number, mergeMethod, false)
 						return prActionDoneMsg{action: "merged", err: err}
 					}
 				} else {
-					m.statusMsg = "⚠️ Press 'm' again to confirm merge (squash)"
+					m.statusMsg = confirmMsg
 					return m, nil
 				}
 			}
@@ -624,12 +352,31 @@ func (m dashModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 				}
 			}
 		case "A":
-			// AI analysis
+			// AI analysis (check cache first)
 			if m.aiAvailable && m.viewMode == viewDetail && m.detailPR != nil {
+				pr := m.detailPR
+				// Check cache first (1 hour TTL)
+				if m.db != nil {
+					cached, err := m.db.GetAIAnalysis(pr.Repo, pr.Number, 1*time.Hour)
+					if err == nil && cached != nil {
+						m.aiAnalysis = &ai.PRAnalysis{
+							Summary:       cached.Summary,
+							ActionNeeded:  cached.ActionNeeded,
+							ReviewSummary: cached.ReviewSummary,
+							RiskLevel:     cached.RiskLevel,
+							BlockedBy:     cached.BlockedBy,
+						}
+						// Parse suggested fixes from JSON
+						if cached.SuggestedFixes != "" {
+							json.Unmarshal([]byte(cached.SuggestedFixes), &m.aiAnalysis.SuggestedFixes)
+						}
+						m.statusMsg = "✓ AI analysis (cached)"
+						return m, nil
+					}
+				}
 				m.aiLoading = true
 				m.aiAnalysis = nil
 				m.aiThread = nil
-				pr := m.detailPR
 				repoPath := m.findLocalRepo(pr.Repo)
 				return m, func() tea.Msg {
 					analysis, err := ai.AnalyzePR(pr.Repo, pr.Number, repoPath)
@@ -798,6 +545,18 @@ func (m dashModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		} else {
 			m.aiAnalysis = msg.analysis
 			m.statusMsg = "✓ AI analysis complete"
+			// Cache the result
+			if m.db != nil && m.detailPR != nil && msg.analysis != nil {
+				fixesJSON, _ := json.Marshal(msg.analysis.SuggestedFixes)
+				m.db.UpsertAIAnalysis(m.detailPR.Repo, m.detailPR.Number, &cache.CachedAIAnalysis{
+					Summary:        msg.analysis.Summary,
+					ActionNeeded:   msg.analysis.ActionNeeded,
+					ReviewSummary:  msg.analysis.ReviewSummary,
+					RiskLevel:      msg.analysis.RiskLevel,
+					SuggestedFixes: string(fixesJSON),
+					BlockedBy:      msg.analysis.BlockedBy,
+				})
+			}
 		}
 
 	case aiThreadDoneMsg:
@@ -838,1201 +597,4 @@ func (m dashModel) currentListLen() int {
 		return len(m.needsAttention)
 	}
 	return 0
-}
-
-func (m *dashModel) openDetail() (tea.Model, tea.Cmd) {
-	var pr *cache.CachedPR
-	switch m.section {
-	case sectionDoNow:
-		if m.cursor < len(m.doNow) {
-			pr = &m.doNow[m.cursor]
-		}
-	case sectionWaiting:
-		if m.cursor < len(m.waiting) {
-			pr = &m.waiting[m.cursor]
-		}
-	case sectionReview:
-		if m.cursor < len(m.review) {
-			pr = &m.review[m.cursor]
-		}
-	case sectionNeedsAttention:
-		if m.cursor < len(m.needsAttention) {
-			pr = &m.needsAttention[m.cursor]
-		}
-	case sectionWorkspace:
-		// Workspace items show path info and open in browser
-		if m.cursor < len(m.workspace) {
-			ws := m.workspace[m.cursor]
-			m.statusMsg = fmt.Sprintf("📁 %s", ws.Path)
-			if ws.HasRemote {
-				gh.OpenInBrowser(fmt.Sprintf("https://github.com/%s", ws.Name))
-			}
-		}
-		return m, nil
-	}
-	if pr == nil {
-		return m, nil
-	}
-
-	return m, loadDetail(pr)
-}
-
-func loadDetail(pr *cache.CachedPR) tea.Cmd {
-	return func() tea.Msg {
-		detail, err := gh.GetPRDetail(pr.Repo, pr.Number)
-		if err != nil {
-			return detailLoadedMsg{err: err}
-		}
-		cached := &cache.CachedPR{PR: *detail, Repo: pr.Repo, Section: pr.Section}
-
-		threads, _ := gh.GetReviewThreads(pr.Repo, pr.Number)
-
-		return detailLoadedMsg{pr: cached, threads: threads}
-	}
-}
-
-func (m *dashModel) openInBrowser() {
-	// In detail view, open the selected thread URL if available
-	if m.viewMode == viewDetail {
-		if m.detailPR != nil {
-			// Try to open the thread URL at the cursor
-			if len(m.detailThreads) > 0 {
-				unresolvedIdx := 0
-				for _, t := range m.detailThreads {
-					if t.IsResolved {
-						continue
-					}
-					if unresolvedIdx == m.threadCursor {
-						if len(t.Comments) > 0 {
-							gh.OpenInBrowser(t.Comments[0].URL)
-							return
-						}
-					}
-					unresolvedIdx++
-				}
-			}
-			// Fallback: open the PR URL
-			gh.OpenInBrowser(m.detailPR.URL)
-		}
-		return
-	}
-
-	switch m.section {
-	case sectionDoNow:
-		if m.cursor < len(m.doNow) {
-			gh.OpenInBrowser(m.doNow[m.cursor].URL)
-		}
-	case sectionWaiting:
-		if m.cursor < len(m.waiting) {
-			gh.OpenInBrowser(m.waiting[m.cursor].URL)
-		}
-	case sectionReview:
-		if m.cursor < len(m.review) {
-			gh.OpenInBrowser(m.review[m.cursor].URL)
-		}
-	case sectionNeedsAttention:
-		if m.cursor < len(m.needsAttention) {
-			gh.OpenInBrowser(m.needsAttention[m.cursor].URL)
-		}
-	case sectionWorkspace:
-		if m.cursor < len(m.workspace) {
-			ws := m.workspace[m.cursor]
-			if ws.HasRemote {
-				gh.OpenInBrowser(fmt.Sprintf("https://github.com/%s", ws.Name))
-			}
-		}
-	}
-}
-
-func (m *dashModel) gitPullCmd() tea.Cmd {
-	if m.cursor >= len(m.workspace) {
-		return nil
-	}
-	ws := m.workspace[m.cursor]
-	return func() tea.Msg {
-		_, err := gitCmd(ws.Path, "pull", "origin", ws.Branch)
-		return gitOpDoneMsg{msg: fmt.Sprintf("pull %s/%s", ws.Name, ws.Branch), err: err}
-	}
-}
-
-func (m *dashModel) gitPushCmd() tea.Cmd {
-	if m.cursor >= len(m.workspace) {
-		return nil
-	}
-	ws := m.workspace[m.cursor]
-	return func() tea.Msg {
-		_, err := gitCmd(ws.Path, "push", "origin", ws.Branch)
-		return gitOpDoneMsg{msg: fmt.Sprintf("push %s/%s", ws.Name, ws.Branch), err: err}
-	}
-}
-
-func (m *dashModel) fetchAllCmd() tea.Cmd {
-	workspace := m.workspace
-	return func() tea.Msg {
-		failed := 0
-		for _, ws := range workspace {
-			_, err := gitCmd(ws.Path, "fetch", "--all")
-			if err != nil {
-				failed++
-			}
-		}
-		msg := fmt.Sprintf("fetched %d repos", len(workspace))
-		var err error
-		if failed > 0 {
-			msg = fmt.Sprintf("fetched %d repos (%d failed)", len(workspace), failed)
-		}
-		return gitOpDoneMsg{msg: msg, err: err}
-	}
-}
-
-// updateSearch handles key input in search mode
-func (m *dashModel) updateSearch(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "ctrl+c":
-		return m, tea.Quit
-	case "esc":
-		m.viewMode = viewList
-		return m, nil
-	case "enter":
-		if len(m.searchResults) > 0 && m.searchCursor < len(m.searchResults) {
-			// Clone selected repo
-			repo := m.searchResults[m.searchCursor]
-			m.statusMsg = fmt.Sprintf("Cloning %s...", repo)
-			return m, m.cloneRepoCmd(repo)
-		}
-		if len(m.searchQuery) > 0 && !m.searching {
-			// Execute search
-			m.searching = true
-			q := m.searchQuery
-			return m, func() tea.Msg {
-				repos, err := gh.SearchOrgRepos(q)
-				return searchResultsMsg{repos: repos, err: err}
-			}
-		}
-	case "up":
-		if m.searchCursor > 0 {
-			m.searchCursor--
-		}
-	case "down":
-		if len(m.searchResults) > 0 && m.searchCursor < len(m.searchResults)-1 {
-			m.searchCursor++
-		}
-	case "backspace":
-		if len(m.searchQuery) > 0 {
-			m.searchQuery = m.searchQuery[:len(m.searchQuery)-1]
-			m.searchResults = nil
-		}
-	default:
-		ch := msg.String()
-		if len(ch) == 1 && ch[0] >= 32 {
-			m.searchQuery += ch
-			m.searchResults = nil
-		}
-	}
-	return m, nil
-}
-
-// updateReply handles key input in reply mode
-func (m *dashModel) updateReply(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "ctrl+c":
-		return m, tea.Quit
-	case "esc":
-		// Cancel reply and return to detail view
-		m.viewMode = viewDetail
-		m.replyText = ""
-		m.replyThreadID = ""
-		return m, nil
-	case "enter":
-		// Submit reply if text is not empty
-		if len(strings.TrimSpace(m.replyText)) > 0 && m.detailPR != nil {
-			pr := m.detailPR
-			threadID := m.replyThreadID
-			replyText := m.replyText
-			
-			// Return to detail view and post reply
-			m.viewMode = viewDetail
-			m.replyText = ""
-			m.replyThreadID = ""
-			m.statusMsg = "Posting reply..."
-			
-			return m, func() tea.Msg {
-				err := gh.ReplyToComment(pr.Repo, pr.Number, threadID, replyText)
-				return prActionDoneMsg{action: "replied", err: err}
-			}
-		}
-		return m, nil
-	case "backspace":
-		if len(m.replyText) > 0 {
-			m.replyText = m.replyText[:len(m.replyText)-1]
-		}
-	case "ctrl+u":
-		// Clear entire input
-		m.replyText = ""
-	default:
-		// Add typed character to reply text
-		ch := msg.String()
-		if len(ch) == 1 && ch[0] >= 32 {
-			m.replyText += ch
-		}
-	}
-	return m, nil
-}
-
-func (m *dashModel) checkoutPRCmd() tea.Cmd {
-	var pr *cache.CachedPR
-	switch m.section {
-	case sectionDoNow:
-		if m.cursor < len(m.doNow) {
-			pr = &m.doNow[m.cursor]
-		}
-	case sectionWaiting:
-		if m.cursor < len(m.waiting) {
-			pr = &m.waiting[m.cursor]
-		}
-	case sectionReview:
-		if m.cursor < len(m.review) {
-			pr = &m.review[m.cursor]
-		}
-	case sectionNeedsAttention:
-		if m.cursor < len(m.needsAttention) {
-			pr = &m.needsAttention[m.cursor]
-		}
-	}
-	if pr == nil {
-		return nil
-	}
-
-	repo := pr.Repo
-	branch := pr.HeadRefName
-	number := pr.Number
-
-	// Capture workspace scan dirs for use inside the closure
-	localPath := m.findLocalRepo(repo)
-	reposDir := m.cfg.Settings.ReposDir
-
-	m.statusMsg = fmt.Sprintf("Checking out PR #%d...", number)
-	return func() tea.Msg {
-		// If branch name is unknown (search results), fetch it first
-		if branch == "" {
-			detail, err := gh.GetPRDetail(repo, number)
-			if err != nil || detail.HeadRefName == "" {
-				return checkoutDoneMsg{repo: repo, branch: "unknown", err: fmt.Errorf("can't determine branch for PR #%d", number)}
-			}
-			branch = detail.HeadRefName
-		}
-
-		// If repo exists locally, checkout there
-		if localPath != "" {
-			gitCmd(localPath, "fetch", "--all")
-			_, err := gitCmd(localPath, "checkout", branch)
-			if err != nil {
-				_, err = gitCmd(localPath, "checkout", "-b", branch, "origin/"+branch)
-			}
-			return checkoutDoneMsg{repo: repo, branch: branch, err: err}
-		}
-
-		// Not found locally — clone first, then checkout
-		dest := reposDir + "/" + repo
-		if !isGitRepo(dest) {
-			err := gh.CloneRepo(repo, dest)
-			if err != nil {
-				return checkoutDoneMsg{repo: repo, branch: branch, err: fmt.Errorf("clone failed: %w", err)}
-			}
-		}
-		gitCmd(dest, "fetch", "--all")
-		_, err := gitCmd(dest, "checkout", branch)
-		if err != nil {
-			_, err = gitCmd(dest, "checkout", "-b", branch, "origin/"+branch)
-		}
-		return checkoutDoneMsg{repo: repo, branch: branch, err: err}
-	}
-}
-
-func (m *dashModel) cloneCurrentPRRepo() tea.Cmd {
-	var repo string
-	switch m.section {
-	case sectionDoNow:
-		if m.cursor < len(m.doNow) {
-			repo = m.doNow[m.cursor].Repo
-		}
-	case sectionWaiting:
-		if m.cursor < len(m.waiting) {
-			repo = m.waiting[m.cursor].Repo
-		}
-	case sectionReview:
-		if m.cursor < len(m.review) {
-			repo = m.review[m.cursor].Repo
-		}
-	case sectionNeedsAttention:
-		if m.cursor < len(m.needsAttention) {
-			repo = m.needsAttention[m.cursor].Repo
-		}
-	}
-	if repo == "" {
-		return nil
-	}
-	return m.cloneRepoCmd(repo)
-}
-
-func (m *dashModel) cloneRepoCmd(repo string) tea.Cmd {
-	// Check if already exists locally
-	localPath := m.findLocalRepo(repo)
-	if localPath != "" {
-		m.statusMsg = fmt.Sprintf("✓ Already cloned at %s", localPath)
-		return nil
-	}
-
-	reposDir := m.cfg.Settings.ReposDir
-	dest := reposDir + "/" + repo
-	m.statusMsg = fmt.Sprintf("Cloning %s...", repo)
-	return func() tea.Msg {
-		// Double-check destination doesn't exist
-		if isGitRepo(dest) {
-			return cloneDoneMsg{repo: repo, path: dest, err: nil}
-		}
-		err := gh.CloneRepo(repo, dest)
-		return cloneDoneMsg{repo: repo, path: dest, err: err}
-	}
-}
-
-// findLocalRepo searches workspace scan dirs for a repo matching the given name
-func (m *dashModel) findLocalRepo(repo string) string {
-	parts := strings.Split(repo, "/")
-	repoName := repo
-	if len(parts) == 2 {
-		repoName = parts[1]
-	}
-
-	// Fast path: check already-scanned workspace results first (no subprocess)
-	for _, ws := range m.workspace {
-		if ws.Name == repo || ws.Name == repoName {
-			return ws.Path
-		}
-	}
-
-	// Check explicit workspace repos mapping
-	if path, ok := m.cfg.Workspace.Repos[repo]; ok {
-		if isGitRepo(path) {
-			return path
-		}
-	}
-
-	// Check in the configured repos dir
-	reposDir := m.cfg.Settings.ReposDir
-	if reposDir != "" {
-		path := reposDir + "/" + repo
-		if isGitRepo(path) {
-			return path
-		}
-		path = reposDir + "/" + repoName
-		if isGitRepo(path) {
-			return path
-		}
-	}
-
-	// Slow path: scan dirs (subprocess per check)
-	for _, dir := range m.cfg.Workspace.ScanDirs {
-		path := dir + "/" + repoName
-		if isGitRepo(path) {
-			return path
-		}
-		if len(parts) == 2 {
-			path = dir + "/" + parts[0] + "/" + parts[1]
-			if isGitRepo(path) {
-				return path
-			}
-		}
-	}
-
-	return ""
-}
-
-// View renders the dashboard
-func (m dashModel) View() string {
-	if m.viewMode == viewSearch {
-		return m.viewSearchMode()
-	}
-	if m.viewMode == viewReply {
-		return m.viewReplyMode()
-	}
-	if m.viewMode == viewDetail && m.detailPR != nil {
-		return m.viewDetail()
-	}
-	return m.viewDashboard()
-}
-
-func (m dashModel) viewSearchMode() string {
-	width := m.width
-	if width < 60 {
-		width = 80
-	}
-
-	var s strings.Builder
-
-	header := headerStyle.Width(width).Render(" 🔍 Search Repos")
-	s.WriteString(header + "\n\n")
-
-	// Search input
-	cursor := "█"
-	if m.searching {
-		cursor = m.spinFrames[m.spinner%len(m.spinFrames)]
-	}
-	s.WriteString(fmt.Sprintf("  Search: %s%s\n\n", m.searchQuery, cursor))
-
-	if m.searching {
-		s.WriteString(fmt.Sprintf("  %s Searching...\n", m.spinFrames[m.spinner%len(m.spinFrames)]))
-	} else if len(m.searchResults) > 0 {
-		s.WriteString(fmt.Sprintf("  %d repos found:\n\n", len(m.searchResults)))
-
-		maxShow := m.height - 10
-		if maxShow < 5 {
-			maxShow = 15
-		}
-
-		for i, repo := range m.searchResults {
-			if i >= maxShow {
-				s.WriteString(fmt.Sprintf("\n  + %d more...\n", len(m.searchResults)-maxShow))
-				break
-			}
-			cursor := "  "
-			if i == m.searchCursor {
-				cursor = prTitleSelectedStyle.Render("▸ ")
-			}
-
-			// Check if repo exists locally
-			localPath := m.findLocalRepo(repo)
-			localTag := ""
-			if localPath != "" {
-				localTag = wsCleanStyle.Render(" (local)")
-			}
-
-			if i == m.searchCursor {
-				s.WriteString(fmt.Sprintf("  %s%s%s\n", cursor, prTitleSelectedStyle.Render(repo), localTag))
-			} else {
-				s.WriteString(fmt.Sprintf("  %s%s%s\n", cursor, repo, localTag))
-			}
-		}
-	} else if m.searchQuery != "" && !m.searching {
-		s.WriteString("  Press [enter] to search\n")
-	} else {
-		s.WriteString("  Type a query to search repos in your orgs.\n")
-		s.WriteString("  Select a result and press [enter] to clone.\n")
-	}
-
-	s.WriteString("\n")
-	s.WriteString(helpStyle.Render("  " + strings.Join([]string{
-		helpPair("enter", "search/clone"),
-		helpPair("↑↓", "nav"),
-		helpPair("esc", "back"),
-	}, "  ")))
-
-	return s.String()
-}
-
-func (m dashModel) viewReplyMode() string {
-	width := m.width
-	if width < 60 {
-		width = 80
-	}
-
-	var s strings.Builder
-
-	// Header
-	header := headerStyle.Width(width).Render(" 💬 Reply to Review Comment")
-	s.WriteString(header + "\n\n")
-
-	// Show current thread context if available
-	if m.detailPR != nil && m.replyThreadID != "" {
-		// Find the thread
-		for _, t := range m.detailThreads {
-			if t.ID == m.replyThreadID && len(t.Comments) > 0 {
-				firstComment := t.Comments[0]
-				s.WriteString(detailLabelStyle.Render("Original Comment") + "\n")
-				s.WriteString(fmt.Sprintf("  %s: %s\n\n", threadAuthorStyle.Render("@"+firstComment.Author), threadBodyStyle.Render(firstComment.Body)))
-				break
-			}
-		}
-	}
-
-	// Reply input box
-	s.WriteString(detailLabelStyle.Render("Your Reply") + "\n")
-	cursor := "█"
-	replyLines := strings.Split(m.replyText, "\n")
-	if len(replyLines) == 0 || m.replyText == "" {
-		s.WriteString(fmt.Sprintf("  %s%s\n", m.replyText, cursor))
-	} else {
-		for i, line := range replyLines {
-			if i == len(replyLines)-1 {
-				s.WriteString(fmt.Sprintf("  %s%s\n", line, cursor))
-			} else {
-				s.WriteString(fmt.Sprintf("  %s\n", line))
-			}
-		}
-	}
-
-	s.WriteString("\n")
-	
-	// Character count
-	charCount := len(m.replyText)
-	countStyle := wsMetaStyle
-	if charCount > 500 {
-		countStyle = wsDirtyStyle
-	}
-	if charCount > 1000 {
-		countStyle = wsBehindStyle
-	}
-	s.WriteString(fmt.Sprintf("  %s\n\n", countStyle.Render(fmt.Sprintf("%d characters", charCount))))
-
-	// Help text
-	s.WriteString(helpStyle.Render("  " + strings.Join([]string{
-		helpPair("enter", "post"),
-		helpPair("ctrl+u", "clear"),
-		helpPair("esc", "cancel"),
-	}, "  ")))
-
-	return s.String()
-}
-
-func (m dashModel) viewDashboard() string {
-	// ─── Header Bar ─────────────────────────────────────────
-	syncAgo := ""
-	if !m.lastSync.IsZero() {
-		syncAgo = fmt.Sprintf("  ⟳ %s", timeSince(m.lastSync))
-	}
-	totalPRs := len(m.doNow) + len(m.waiting) + len(m.review)
-
-	headerWidth := m.width
-	if headerWidth < 60 {
-		headerWidth = 80
-	}
-	header := headerStyle.Width(headerWidth).Render(
-		fmt.Sprintf(" ⚡ PRFlow    @%s  ·  %d active PRs%s", m.username, totalPRs, syncAgo))
-
-	// ─── Sidebar ────────────────────────────────────────────
-	sidebar := m.renderSidebar()
-
-	// ─── Main Panel ─────────────────────────────────────────
-	mainWidth := m.width - 28
-	if mainWidth < 40 {
-		mainWidth = 60
-	}
-	main := m.renderMainPanel(mainWidth)
-
-	// ─── Status Bar ─────────────────────────────────────────
-	statusBar := m.renderStatusBar()
-
-	// ─── Help Bar ───────────────────────────────────────────
-	help := m.renderHelp()
-
-	// ─── Compose Layout ─────────────────────────────────────
-	content := lipgloss.JoinHorizontal(lipgloss.Top, sidebar, main)
-
-	return header + "\n" + content + "\n" + statusBar + help
-}
-
-func (m dashModel) renderSidebar() string {
-	var s strings.Builder
-
-	type sidebarEntry struct {
-		icon  string
-		label string
-		count int
-		sec   section
-	}
-
-	entries := []sidebarEntry{
-		{"⚡", "Do Now", len(m.doNow), sectionDoNow},
-		{"⏳", "Waiting", len(m.waiting), sectionWaiting},
-		{"👀", "Review", len(m.review), sectionReview},
-		{"📂", "Workspace", len(m.workspace), sectionWorkspace},
-		{"🔔", "Needs Attention", len(m.needsAttention), sectionNeedsAttention},
-	}
-
-	for _, e := range entries {
-		countStr := sidebarCountStyle.Render(fmt.Sprintf(" %d", e.count))
-		if e.sec == m.section {
-			s.WriteString(sidebarActiveStyle.Render(
-				fmt.Sprintf("▸ %s %s", e.icon, e.label)) + countStr + "\n")
-		} else {
-			s.WriteString(sidebarSectionStyle.Render(
-				fmt.Sprintf("  %s %s", e.icon, e.label)) + countStr + "\n")
-		}
-	}
-
-	// Favorites
-	if len(m.cfg.Favorites) > 0 {
-		s.WriteString("\n")
-		s.WriteString(favHeaderStyle.Render("★ Favorites") + "\n")
-		for _, fav := range m.cfg.Favorites {
-			parts := strings.Split(fav, "/")
-			name := fav
-			if len(parts) == 2 {
-				name = parts[1]
-			}
-			s.WriteString(favItemStyle.Render(name) + "\n")
-		}
-	}
-
-	return sidebarStyle.Render(s.String())
-}
-
-func (m dashModel) renderMainPanel(width int) string {
-	var s strings.Builder
-
-	// Section header with rule
-	s.WriteString(sectionHeader.Width(width - 4).Render(m.section.String()))
-	s.WriteString("\n")
-
-	if m.loading {
-		spin := m.spinFrames[m.spinner%len(m.spinFrames)]
-		s.WriteString(fmt.Sprintf("\n %s Syncing with GitHub...\n", spin))
-		return mainPanelStyle.Width(width).Render(s.String())
-	}
-
-	switch m.section {
-	case sectionDoNow:
-		s.WriteString(m.renderPRCards(m.doNow, width))
-	case sectionWaiting:
-		s.WriteString(m.renderPRCards(m.waiting, width))
-	case sectionReview:
-		s.WriteString(m.renderPRCards(m.review, width))
-	case sectionWorkspace:
-		s.WriteString(m.renderWorkspaceCards(width))
-	case sectionNeedsAttention:
-		s.WriteString(m.renderPRCards(m.needsAttention, width))
-	}
-
-	return mainPanelStyle.Width(width).Render(s.String())
-}
-
-func (m dashModel) renderPRCards(prs []cache.CachedPR, width int) string {
-	if len(prs) == 0 {
-		return emptyStyle.Render("Nothing here — you're all caught up! 🎉")
-	}
-
-	var s strings.Builder
-	maxShow := m.height - 10
-	if maxShow < 5 {
-		maxShow = 15
-	}
-
-	// Scroll window: keep cursor visible
-	start := 0
-	if m.cursor >= maxShow {
-		start = m.cursor - maxShow + 1
-	}
-
-	for i := start; i < len(prs); i++ {
-		if i-start >= maxShow {
-			s.WriteString(fmt.Sprintf("\n  %s\n",
-				repoStyle.Render(fmt.Sprintf("+ %d more...", len(prs)-i))))
-			break
-		}
-		s.WriteString(m.renderPRCard(prs[i], i == m.cursor, width-6))
-	}
-	return s.String()
-}
-
-func (m dashModel) renderPRCard(pr cache.CachedPR, selected bool, width int) string {
-	// Repo short name
-	parts := strings.Split(pr.Repo, "/")
-	repoShort := pr.Repo
-	if len(parts) == 2 {
-		repoShort = parts[1]
-	}
-
-	// Line 1: repo  #number  title
-	numStr := prNumberStyle.Render(fmt.Sprintf("#%d", pr.Number))
-	repoStr := prRepoStyle.Render(repoShort)
-
-	maxTitle := width - len(repoShort) - 10
-	if maxTitle < 20 {
-		maxTitle = 30
-	}
-	title := truncate(pr.Title, maxTitle)
-
-	titleStr := prTitleStyle.Render(title)
-	if selected {
-		titleStr = prTitleSelectedStyle.Render(title)
-	}
-
-	line1 := fmt.Sprintf("%s  %s  %s", repoStr, numStr, titleStr)
-
-	// Line 2: status badge + time + stale reviewer indicator
-	badge := m.prBadge(pr)
-	timeAgo := ""
-	if pr.UpdatedAt != "" {
-		timeAgo = wsMetaStyle.Render(fmt.Sprintf("  updated %s", formatTimeAgo(pr.UpdatedAt)))
-	}
-	
-	// Add stale reviewer indicator (if any reviewer waiting 3+ days)
-	staleIndicator := ""
-	reviewerStatuses := CalculateReviewerStatus(&pr.PR)
-	if len(reviewerStatuses) > 0 {
-		stalest := reviewerStatuses[0] // already sorted by wait time descending
-		if stalest.WaitDays >= 3 {
-			staleIndicator = wsBehindStyle.Render(fmt.Sprintf("  ⏰ @%s (%dd)", stalest.Login, stalest.WaitDays))
-		}
-	}
-	
-	line2 := badge + timeAgo + staleIndicator
-
-	content := line1 + "\n" + line2
-
-	if selected {
-		return prCardSelectedStyle.Width(width).Render(content)
-	}
-	return prCardStyle.Width(width).Render(content)
-}
-
-func (m dashModel) prBadge(pr cache.CachedPR) string {
-	switch {
-	case pr.ReviewDecision == "APPROVED" && pr.Mergeable != "CONFLICTING":
-		// Check CI status for approved PRs
-		if HasPassingCI(pr) {
-			return badgeMerge.Render("READY ✓CI")
-		} else {
-			// CI failing or pending
-			hasFailedCI := false
-			for _, check := range pr.StatusCheckRollup {
-				if check.Conclusion == "FAILURE" || check.Conclusion == "TIMED_OUT" || check.Conclusion == "ACTION_REQUIRED" {
-					hasFailedCI = true
-					break
-				}
-			}
-			if hasFailedCI {
-				return badgeConflict.Render("READY ✗CI")
-			}
-			// CI pending/in-progress
-			return badgeWaiting.Render("READY ⏳CI")
-		}
-	case pr.ReviewDecision == "APPROVED" && pr.Mergeable == "CONFLICTING":
-		return badgeConflict.Render("CONFLICT")
-	case pr.ReviewDecision == "CHANGES_REQUESTED":
-		return badgeChanges.Render("CHANGES REQUESTED")
-	case pr.Mergeable == "CONFLICTING":
-		return badgeConflict.Render("CONFLICT")
-	case pr.IsDraft:
-		return badgeDraft.Render("DRAFT")
-	case pr.ReviewDecision == "REVIEW_REQUIRED":
-		return badgeWaiting.Render("AWAITING REVIEW")
-	default:
-		return badgeWaiting.Render("IN REVIEW")
-	}
-}
-
-func (m dashModel) renderWorkspaceCards(width int) string {
-	if len(m.workspace) == 0 {
-		return emptyStyle.Render("No repos found.\nConfigure workspace.scan_dirs in ~/.config/prflow/config.yaml")
-	}
-
-	var s strings.Builder
-	maxShow := m.height - 10
-	if maxShow < 5 {
-		maxShow = 15
-	}
-
-	// Scroll window: keep cursor visible
-	start := 0
-	if m.cursor >= maxShow {
-		start = m.cursor - maxShow + 1
-	}
-
-	for i := start; i < len(m.workspace); i++ {
-		if i-start >= maxShow {
-			s.WriteString(fmt.Sprintf("\n  %s\n",
-				repoStyle.Render(fmt.Sprintf("+ %d more...", len(m.workspace)-i))))
-			break
-		}
-		ws := m.workspace[i]
-		s.WriteString(m.renderWorkspaceCard(&ws, i == m.cursor, width-6))
-	}
-	return s.String()
-}
-
-func (m dashModel) renderWorkspaceCard(ws *RepoStatus, selected bool, width int) string {
-	nameStr := prNumberStyle.Render(ws.Name)
-
-	// Branch
-	branchStr := wsMetaStyle.Render("on ") + detailValueStyle.Render(ws.Branch)
-
-	// Behind/ahead
-	var baStr string
-	if ws.Behind > 0 && ws.Behind > 20 {
-		baStr = wsBehindStyle.Render(fmt.Sprintf("↓%d behind", ws.Behind)) + "  "
-	} else if ws.Behind > 0 {
-		baStr = wsDirtyStyle.Render(fmt.Sprintf("↓%d behind", ws.Behind)) + "  "
-	}
-	if ws.Ahead > 0 {
-		baStr += wsCleanStyle.Render(fmt.Sprintf("↑%d ahead", ws.Ahead))
-	}
-	if ws.Behind == 0 && ws.Ahead == 0 {
-		baStr = wsCleanStyle.Render("up to date")
-	}
-
-	// Working tree
-	var treeStr string
-	if ws.Clean {
-		treeStr = wsCleanStyle.Render("✓ clean")
-	} else {
-		var parts []string
-		if ws.Modified > 0 {
-			parts = append(parts, fmt.Sprintf("%d modified", ws.Modified))
-		}
-		if ws.Staged > 0 {
-			parts = append(parts, fmt.Sprintf("%d staged", ws.Staged))
-		}
-		if ws.Untracked > 0 {
-			parts = append(parts, fmt.Sprintf("%d untracked", ws.Untracked))
-		}
-		treeStr = wsDirtyStyle.Render(strings.Join(parts, " · "))
-	}
-
-	// Unpushed
-	unpushedStr := ""
-	if ws.Unpushed > 0 {
-		unpushedStr = "\n" + wsDirtyStyle.Render(fmt.Sprintf("⬆ %d unpushed", ws.Unpushed))
-	}
-
-	// Last commit
-	commitStr := ""
-	if ws.LastCommit != "" {
-		commitStr = "\n" + wsMetaStyle.Render(ws.LastCommit)
-	}
-
-	content := nameStr + "  " + branchStr + "\n" + baStr + "  " + treeStr + unpushedStr + commitStr
-
-	if selected {
-		return wsCardSelectedStyle.Width(width).Render(content)
-	}
-	return wsCardStyle.Width(width).Render(content)
-}
-
-func (m dashModel) viewDetail() string {
-	pr := m.detailPR
-	var s strings.Builder
-
-	width := m.width
-	if width < 60 {
-		width = 80
-	}
-
-	// Header
-	header := headerStyle.Width(width).Render(
-		fmt.Sprintf(" %s  #%d", pr.Repo, pr.Number))
-	s.WriteString(header + "\n\n")
-
-	// Title
-	s.WriteString("  " + lipgloss.NewStyle().Bold(true).Foreground(colorWhite).Render(pr.Title) + "\n\n")
-
-	// Info grid
-	s.WriteString(fmt.Sprintf("  %s %s → %s\n",
-		detailLabelStyle.Render("Branch"),
-		detailValueStyle.Render(pr.HeadRefName),
-		wsMetaStyle.Render(pr.BaseRefName)))
-
-	// Status badge
-	s.WriteString(fmt.Sprintf("  %s %s\n", detailLabelStyle.Render("Status"), m.prBadge(cache.CachedPR{PR: pr.PR})))
-
-	// Mergeable
-	if pr.Mergeable != "" {
-		mergeIcon := "  "
-		if pr.Mergeable == "MERGEABLE" {
-			mergeIcon = wsCleanStyle.Render("✓ MERGEABLE")
-		} else if pr.Mergeable == "CONFLICTING" {
-			mergeIcon = wsBehindStyle.Render("✗ CONFLICTING")
-		} else {
-			mergeIcon = wsMetaStyle.Render(pr.Mergeable)
-		}
-		s.WriteString(fmt.Sprintf("  %s %s\n", detailLabelStyle.Render("Merge"), mergeIcon))
-	}
-
-	// CI Checks
-	if len(pr.StatusCheckRollup) > 0 {
-		s.WriteString(fmt.Sprintf("\n  %s\n", detailLabelStyle.Render("CI Checks")))
-		for _, check := range pr.StatusCheckRollup {
-			var icon string
-			switch check.Conclusion {
-			case "SUCCESS":
-				icon = wsCleanStyle.Render("✓")
-			case "FAILURE":
-				icon = wsBehindStyle.Render("✗")
-			default:
-				icon = wsDirtyStyle.Render("⏳")
-			}
-			s.WriteString(fmt.Sprintf("    %s %s\n", icon, check.Name))
-		}
-	}
-
-	// Reviewers with wait times
-	reviewerStatuses := CalculateReviewerStatus(&pr.PR)
-	if len(reviewerStatuses) > 0 {
-		s.WriteString(fmt.Sprintf("\n  %s\n", detailLabelStyle.Render("Reviewers")))
-		for _, status := range reviewerStatuses {
-			icon := reviewerIcon(status.State)
-			nameStr := threadAuthorStyle.Render("@" + status.Login)
-			stateStr := formatReviewState(status.State)
-			waitStr := formatWaitTime(status.WaitDays, status.IsPending)
-			s.WriteString(fmt.Sprintf("    %s %s  %s  %s\n", icon, nameStr, stateStr, waitStr))
-		}
-	}
-
-	// Review Threads
-	if len(m.detailThreads) > 0 {
-		unresolved := 0
-		resolved := 0
-		for _, t := range m.detailThreads {
-			if t.IsResolved {
-				resolved++
-			} else {
-				unresolved++
-			}
-		}
-
-		s.WriteString("\n" + threadHeaderStyle.Render(
-			fmt.Sprintf("  📝 Unresolved Threads (%d)", unresolved)) + "\n\n")
-
-		threadIdx := 0
-		for _, t := range m.detailThreads {
-			if t.IsResolved {
-				continue
-			}
-			selected := threadIdx == m.threadCursor
-
-			fileStr := threadFileStyle.Render(fmt.Sprintf("%s:%d", t.Path, t.Line))
-			var body string
-			if len(t.Comments) > 0 {
-				last := t.Comments[len(t.Comments)-1]
-				body = threadAuthorStyle.Render("@"+last.Author) + "  " +
-					threadBodyStyle.Render(truncate(last.Body, 70))
-			}
-
-			content := fileStr + "\n" + body
-			cardWidth := width - 8
-			if cardWidth < 40 {
-				cardWidth = 60
-			}
-
-			if selected {
-				s.WriteString("  " + threadCardSelectedStyle.Width(cardWidth).Render(content) + "\n")
-			} else {
-				s.WriteString("  " + threadCardStyle.Width(cardWidth).Render(content) + "\n")
-			}
-			threadIdx++
-		}
-
-		if resolved > 0 {
-			s.WriteString(fmt.Sprintf("  %s\n",
-				wsMetaStyle.Render(fmt.Sprintf("  ✅ %d resolved threads (collapsed)", resolved))))
-		}
-	}
-
-	// AI Analysis (if available)
-	if m.aiLoading {
-		s.WriteString(fmt.Sprintf("\n  %s %s Analyzing with Claude Code...\n",
-			threadHeaderStyle.Render("🤖 AI Analysis"),
-			m.spinFrames[m.spinner%len(m.spinFrames)]))
-	} else if m.aiAnalysis != nil {
-		a := m.aiAnalysis
-		s.WriteString("\n" + threadHeaderStyle.Render("  🤖 AI Analysis") + "\n\n")
-
-		if a.Summary != "" {
-			s.WriteString(fmt.Sprintf("  %s %s\n", detailLabelStyle.Render("Summary"), detailValueStyle.Render(a.Summary)))
-		}
-		if a.ActionNeeded != "" {
-			s.WriteString(fmt.Sprintf("  %s %s\n", detailLabelStyle.Render("Next Action"), wsCleanStyle.Render(a.ActionNeeded)))
-		}
-		if a.ReviewSummary != "" {
-			s.WriteString(fmt.Sprintf("  %s %s\n", detailLabelStyle.Render("Reviews"), detailValueStyle.Render(a.ReviewSummary)))
-		}
-		if a.RiskLevel != "" {
-			riskStyle := wsCleanStyle
-			if a.RiskLevel == "medium" {
-				riskStyle = wsDirtyStyle
-			} else if a.RiskLevel == "high" {
-				riskStyle = wsBehindStyle
-			}
-			s.WriteString(fmt.Sprintf("  %s %s\n", detailLabelStyle.Render("Risk"), riskStyle.Render(a.RiskLevel)))
-		}
-		if a.BlockedBy != "" {
-			s.WriteString(fmt.Sprintf("  %s %s\n", detailLabelStyle.Render("Blocked By"), wsBehindStyle.Render(a.BlockedBy)))
-		}
-		if len(a.SuggestedFixes) > 0 {
-			s.WriteString(fmt.Sprintf("  %s\n", detailLabelStyle.Render("Suggestions")))
-			for _, fix := range a.SuggestedFixes {
-				s.WriteString(fmt.Sprintf("    → %s\n", detailValueStyle.Render(fix)))
-			}
-		}
-	}
-
-	// AI Thread Analysis (if available)
-	if m.aiThread != nil {
-		t := m.aiThread
-		s.WriteString("\n" + threadHeaderStyle.Render("  🤖 Thread Analysis") + "\n\n")
-		if t.Intent != "" {
-			s.WriteString(fmt.Sprintf("  %s %s\n", detailLabelStyle.Render("Intent"), detailValueStyle.Render(t.Intent)))
-		}
-		if t.Complexity != "" {
-			s.WriteString(fmt.Sprintf("  %s %s\n", detailLabelStyle.Render("Complexity"), detailValueStyle.Render(t.Complexity)))
-		}
-		if t.Suggestion != "" {
-			s.WriteString(fmt.Sprintf("  %s %s\n", detailLabelStyle.Render("Approach"), wsCleanStyle.Render(t.Suggestion)))
-		}
-		if t.DraftReply != "" {
-			s.WriteString(fmt.Sprintf("\n  %s\n", detailLabelStyle.Render("Draft Reply")))
-			s.WriteString(fmt.Sprintf("  %s\n", threadBodyStyle.Render(t.DraftReply)))
-		}
-	}
-
-	// URL
-	s.WriteString(fmt.Sprintf("\n  %s %s\n", detailLabelStyle.Render("URL"), urlStyle.Render(pr.URL)))
-
-	// Help
-	s.WriteString("\n" + m.renderDetailHelp())
-
-	return s.String()
-}
-
-func (m dashModel) renderStatusBar() string {
-	var parts []string
-	if m.statusMsg != "" {
-		parts = append(parts, statusBarStyle.Render("  "+m.statusMsg))
-	}
-	if m.err != "" {
-		parts = append(parts, statusErrorStyle.Render("  ⚠ "+m.err))
-	}
-	if len(parts) == 0 {
-		return ""
-	}
-	return strings.Join(parts, "  ") + "\n"
-}
-
-func (m dashModel) renderHelp() string {
-	var pairs []string
-
-	pairs = append(pairs, helpPair("↑↓", "nav"))
-	pairs = append(pairs, helpPair("tab", "section"))
-	pairs = append(pairs, helpPair("enter", "expand"))
-	pairs = append(pairs, helpPair("o", "open"))
-
-	if m.section == sectionWorkspace {
-		pairs = append(pairs, helpPair("p", "pull"))
-		pairs = append(pairs, helpPair("P", "push"))
-		pairs = append(pairs, helpPair("f", "fetch"))
-	} else {
-		pairs = append(pairs, helpPair("c", "checkout"))
-		pairs = append(pairs, helpPair("C", "clone"))
-	}
-
-	pairs = append(pairs, helpPair("/", "search"))
-	pairs = append(pairs, helpPair("R", "refresh"))
-	pairs = append(pairs, helpPair("q", "quit"))
-
-	return helpStyle.Render("  " + strings.Join(pairs, "  "))
-}
-
-func (m dashModel) renderDetailHelp() string {
-	pairs := []string{
-		helpPair("o", "open in browser"),
-		helpPair("c", "checkout"),
-		helpPair("a", "approve"),
-		helpPair("m", "merge"),
-	}
-	
-	// Only show resolve if there are unresolved threads
-	if len(m.detailThreads) > 0 {
-		hasUnresolved := false
-		for _, t := range m.detailThreads {
-			if !t.IsResolved {
-				hasUnresolved = true
-				break
-			}
-		}
-		if hasUnresolved {
-			pairs = append(pairs, helpPair("r", "resolve"))
-			pairs = append(pairs, helpPair("R", "reply"))
-		}
-	}
-	
-	if m.aiAvailable {
-		pairs = append(pairs, helpPair("A", "AI analyze"))
-	}
-	pairs = append(pairs,
-		helpPair("esc", "back"),
-		helpPair("q", "quit"),
-	)
-	return helpStyle.Render("  " + strings.Join(pairs, "  "))
-}
-
-func timeSince(t time.Time) string {
-	d := time.Since(t)
-	if d < 0 {
-		return "just now"
-	}
-	if d < time.Minute {
-		return "just now"
-	}
-	if d < time.Hour {
-		return fmt.Sprintf("%dm ago", int(d.Minutes()))
-	}
-	hours := int(d.Hours())
-	if hours < 24 {
-		return fmt.Sprintf("%dh ago", hours)
-	}
-	days := hours / 24
-	if days == 1 {
-		return "yesterday"
-	}
-	if days < 30 {
-		return fmt.Sprintf("%dd ago", days)
-	}
-	months := days / 30
-	if months == 1 {
-		return "1 month ago"
-	}
-	if months < 12 {
-		return fmt.Sprintf("%d months ago", months)
-	}
-	years := months / 12
-	if years == 1 {
-		return "1 year ago"
-	}
-	return fmt.Sprintf("%d years ago", years)
-}
-
-func formatTimeAgo(isoTime string) string {
-	if isoTime == "" {
-		return ""
-	}
-	// Try multiple time formats
-	formats := []string{
-		time.RFC3339,
-		"2006-01-02T15:04:05Z",
-		"2006-01-02T15:04:05-07:00",
-		"2006-01-02 15:04:05 -0700 MST",
-		time.RFC3339Nano,
-	}
-	for _, layout := range formats {
-		t, err := time.Parse(layout, isoTime)
-		if err == nil {
-			return timeSince(t)
-		}
-	}
-	return ""
-}
-
-func truncate(s string, max int) string {
-	s = strings.ReplaceAll(s, "\n", " ")
-	s = strings.ReplaceAll(s, "\r", "")
-	if max < 4 {
-		if len(s) > max {
-			return s[:max]
-		}
-		return s
-	}
-	if len(s) > max {
-		return s[:max-3] + "..."
-	}
-	return s
 }
