+++ internal/tui/sorting.go
@@ -6,8 +6,17 @@ import (
 	"github.com/cheenu1092-oss/prflow/internal/cache"
 )
 
-// UrgencyScore calculates a priority score for a PR (higher = more urgent)
-func UrgencyScore(pr cache.CachedPR) int {
+// DefaultStaleThresholdDays is used when no config is available
+const DefaultStaleThresholdDays = 3
+
+// UrgencyScore calculates a priority score for a PR (higher = more urgent).
+// staleThresholdDays controls when a reviewer wait becomes "stale" (default: 3).
+func UrgencyScore(pr cache.CachedPR, staleThresholdDays ...int) int {
+	threshold := DefaultStaleThresholdDays
+	if len(staleThresholdDays) > 0 && staleThresholdDays[0] > 0 {
+		threshold = staleThresholdDays[0]
+	}
+
 	score := 0
 
 	// Critical: Merge conflicts (highest urgency)
@@ -23,24 +32,24 @@ func UrgencyScore(pr cache.CachedPR) int {
 	// Medium-High: Approved and ready to merge (quick win)
 	if pr.ReviewDecision == "APPROVED" && pr.Mergeable != "CONFLICTING" {
 		score += 600
-		
+
 		// Bonus: CI passing (truly ready)
 		if HasPassingCI(pr) {
 			score += 100
 		}
 	}
 
-	// Medium: Waiting on reviewers (time-based urgency)
+	// Medium: Waiting on reviewers (time-based urgency relative to threshold)
 	reviewerStatuses := CalculateReviewerStatus(&pr.PR)
 	if len(reviewerStatuses) > 0 {
 		stalest := reviewerStatuses[0] // already sorted by wait time descending
-		
-		// Stale reviewers (3+ days) add urgency
-		if stalest.WaitDays >= 7 {
-			score += 400 // 1 week = very urgent
-		} else if stalest.WaitDays >= 5 {
+
+		// Scale urgency relative to configured threshold
+		if stalest.WaitDays >= threshold*2+1 {
+			score += 400 // 2x+ threshold = very urgent
+		} else if stalest.WaitDays >= threshold+2 {
 			score += 300
-		} else if stalest.WaitDays >= 3 {
+		} else if stalest.WaitDays >= threshold {
 			score += 200
 		} else if stalest.WaitDays >= 1 {
 			score += 100
@@ -85,17 +94,22 @@ func DaysSinceUpdate(updatedAt string) int {
 	return daysSince(t)
 }
 
-// SortByUrgency sorts a slice of PRs by urgency score (highest first)
-func SortByUrgency(prs []cache.CachedPR) {
+// SortByUrgency sorts a slice of PRs by urgency score (highest first).
+// staleThresholdDays controls when a reviewer wait becomes "stale" (default: 3).
+func SortByUrgency(prs []cache.CachedPR, staleThresholdDays ...int) {
+	threshold := DefaultStaleThresholdDays
+	if len(staleThresholdDays) > 0 && staleThresholdDays[0] > 0 {
+		threshold = staleThresholdDays[0]
+	}
 	sort.SliceStable(prs, func(i, j int) bool {
-		scoreI := UrgencyScore(prs[i])
-		scoreJ := UrgencyScore(prs[j])
-		
+		scoreI := UrgencyScore(prs[i], threshold)
+		scoreJ := UrgencyScore(prs[j], threshold)
+
 		// Higher score = more urgent = comes first
 		if scoreI != scoreJ {
 			return scoreI > scoreJ
 		}
-		
+
 		// Tie-breaker: more recent updates first
 		return prs[i].UpdatedAt > prs[j].UpdatedAt
 	})
